refactor: 将不必要的 long 类型改为 int
游戏数值不会超过 int 范围，long 是过度设计：
- Score: 单局最多百万分，远小于 int 上限 21 亿
- MoveCount: 单局最多几十步
- TileId/ProjectileId: 单局最多几千个
- Tick: 每秒 60 tick，int 可用 414 天

改动范围：
- GameState: Score, MoveCount, NextTileId
- Tile/Projectile: Id, TargetTileId
- RenderCommand: 所有 TileId, ProjectileId
- GameEvent: Tick
- 相关接口和实现的参数签名

保留 long 的场景：
- Analysis 服务的累加器（跨百万局累计）
- DifficultyAnalysis.MaxScorePotential
- 系统 API (GC.GetTotalMemory, 文件大小)

+++ src/Match3.Core.Tests/Systems/Core/AsyncGameLoopTests.cs
@@ -51,7 +51,7 @@ public class AsyncGameLoopTests
             return 0;
         }
 
-        public int ProcessMatches(ref GameState state, List<MatchGroup> groups, long tick, float simTime, IEventCollector events)
+        public int ProcessMatches(ref GameState state, List<MatchGroup> groups, int tick, float simTime, IEventCollector events)
         {
             ProcessMatchesCalled = true;
             return 0;
@@ -61,11 +61,11 @@ public class AsyncGameLoopTests
     private class StubPowerUp : IPowerUpHandler
     {
         public void ActivateBomb(ref GameState state, Position p) { }
-        public void ActivateBomb(ref GameState state, Position p, long tick, float simTime, IEventCollector events) { }
+        public void ActivateBomb(ref GameState state, Position p, int tick, float simTime, IEventCollector events) { }
         public void HandlePowerUp(ref GameState state, Position p, BombType bomb) { }
         public bool TryActivate(ref GameState state, Position p) => false;
         public void ProcessSpecialMove(ref GameState state, Position a, Position b, out int score) { score = 0; }
-        public void ProcessSpecialMove(ref GameState state, Position a, Position b, long tick, float simTime, IEventCollector events, out int score) { score = 0; }
+        public void ProcessSpecialMove(ref GameState state, Position a, Position b, int tick, float simTime, IEventCollector events, out int score) { score = 0; }
     }
 
     private class StubRandom : IRandom
@@ -225,7 +225,7 @@ public class AsyncGameLoopTests
             return groups.Count;
         }
 
-        public int ProcessMatches(ref GameState state, List<MatchGroup> groups, long tick, float simTime, IEventCollector events)
+        public int ProcessMatches(ref GameState state, List<MatchGroup> groups, int tick, float simTime, IEventCollector events)
         {
             ProcessMatchesCalled = true;
             GroupsProcessed = groups.Count;
@@ -314,7 +314,7 @@ public class AsyncGameLoopTests
             LastActivatedPosition = p;
         }
 
-        public void ActivateBomb(ref GameState state, Position p, long tick, float simTime, IEventCollector events)
+        public void ActivateBomb(ref GameState state, Position p, int tick, float simTime, IEventCollector events)
         {
             ActivateBombCalled = true;
             LastActivatedPosition = p;
@@ -323,7 +323,7 @@ public class AsyncGameLoopTests
         public void HandlePowerUp(ref GameState state, Position p, BombType bomb) { }
         public bool TryActivate(ref GameState state, Position p) => false;
         public void ProcessSpecialMove(ref GameState state, Position a, Position b, out int score) { score = 0; }
-        public void ProcessSpecialMove(ref GameState state, Position a, Position b, long tick, float simTime, IEventCollector events, out int score) { score = 0; }
+        public void ProcessSpecialMove(ref GameState state, Position a, Position b, int tick, float simTime, IEventCollector events, out int score) { score = 0; }
     }
 
     #endregion
@@ -370,7 +370,7 @@ public class AsyncGameLoopTests
             return groups.Count;
         }
 
-        public int ProcessMatches(ref GameState state, List<MatchGroup> groups, long tick, float simTime, IEventCollector events)
+        public int ProcessMatches(ref GameState state, List<MatchGroup> groups, int tick, float simTime, IEventCollector events)
         {
             FrameCount++;
             return groups.Count;
@@ -485,7 +485,7 @@ public class AsyncGameLoopTests
     private static void FillBoard(ref GameState state)
     {
         var types = new[] { TileType.Red, TileType.Blue, TileType.Green, TileType.Yellow };
-        long id = 1;
+        int id = 1;
         for (int y = 0; y < state.Height; y++)
         {
             for (int x = 0; x < state.Width; x++)
