+++ src/Match3.Core/Systems/Layers/GroundSystem.cs
@@ -18,7 +18,7 @@ public class GroundSystem : IGroundSystem
         _objectiveSystem = objectiveSystem;
     }
     /// <inheritdoc />
-    public void OnTileDestroyed(ref GameState state, Position position, long tick, float simTime, IEventCollector events)
+    public void OnTileDestroyed(ref GameState state, Position position, int tick, float simTime, IEventCollector events)
     {
         if (!state.IsValid(position))
             return;
