+++ src/Match3.Core/Systems/Swap/ISwapContext.cs
@@ -33,7 +33,7 @@ public interface ISwapContext
         in GameState state,
         Position from,
         Position to,
-        long tick,
+        int tick,
         float simTime,
         IEventCollector events);
 }
